refactor(deep): extract out-of-scope issue-routing helpers to scope_issues.py (#338)
Pure relocation of the #336 scope machinery (fingerprint -> marker ->
cross-run dedup -> filing + post-fix residual revert net) out of the
3546-line orchestrator.py into daydream/deep/scope_issues.py. No behavior
change; _step_fix/_step_fix_gate import the 3 names they call. Only the
dedup test's helper import moves with the module.

Entire-Checkpoint: 9a22393bcda1
